requirements.txt
telebot==0.0.5
pyTelegramBotAPI==4.14.0
langchain==0.1.1
langchain-community==0.0.13
gigachat==0.1.14
python-dotenv==1.0.0
httpx==0.26.0
warnings==0.1.0
smtplib
email
datetime
